Move Request Instance away from the cors related acceptance tests
